Separate data utils from other utils
