+++ pages/index.js
@@ -2,8 +2,8 @@ import BlogCard from '../components/BlogCard'
 import TalkCard from '../components/TalkCard'
 import Layout from '../components/Layout'
 import { METADATA } from '../utils/constants'
-import { getMostRecentPosts } from '../utils/posts'
 import { defaultSocialImage } from '../utils/preview-cards'
+import { getMostRecentPosts } from '../utils/data/posts'
 
 import blogStyles from '../components/BlogCard.module.css'
 import talkStyles from '../components/TalkCard.module.css'
