+++ utils/constants.js
@@ -32,3 +32,7 @@ export const THEME = {
   LIGHT: 'light',
   DARK: 'dark',
 }
+
+// Full length for a single TIL:
+//    https://raw.githubusercontent.com/kale-stew/til/main/<category>/<title>.md
+export const TIL_URL = 'https://raw.githubusercontent.com/kale-stew/til/main'
